Add offline PDB cache and CLI fallback
