+++ metalCoord/cli/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Command handlers for the MetalCoord CLI."""
